+++ src/extension.ts
@@ -115,6 +115,9 @@ export async function activate(context: vscode.ExtensionContext) {
 	vscode.commands.registerTextEditorCommand('zhihu.publish', (textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit) => {
 		publishService.publish(textEditor, edit);
 	})
+	vscode.commands.registerTextEditorCommand('zhihu.publishCurrentMarkdown', (textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit) => {
+		publishService.publish(textEditor, edit);
+	})
 	vscode.commands.registerTextEditorCommand('zhihu.preview', (textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit) => {
 		publishService.preview(textEditor, edit);
 	})